move to package folder, start python3 conversion
