--- cost_calculator/__init__.py
@@ -1,1 +1,0 @@
-__author__ = 'Olivier Van Cutsem'
