+++ electricitycostcalculator/__init__.py
@@ -1,0 +1,4 @@
+__author__ = 'Olivier Van Cutsem'
+
+from electricitycostcalculator import cost_calculator, oadr_signal, openei_tariff
+__all__ = ["cost_calculator", "oadr_signal", "openei_tariff"]
