+++ electricitycostcalculator/oadr_signal/settings_template.json
