+++ electricitycostcalculator/oadr_signal/utils.py
